requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Agent Orchestration
celery==5.3.4
redis==5.0.1

# YouTube & Web Scraping
yt-dlp==2023.11.16
playwright==1.40.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Video Processing
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
pillow==10.1.0
moviepy==1.0.3

# Audio & Transcription
openai-whisper==20231117
whisperx==3.1.1
assemblyai==0.28.0

# AI & ML
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
langchain-openai==0.0.2
llama-index==0.9.13
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
clip-by-openai==1.0
rembg==2.0.50

# Computer Vision
segment-anything==1.0
ultralytics==8.0.206

# Database & Storage
supabase==2.0.3
boto3==1.29.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6

# Publishing APIs
google-auth==2.24.0
tiktok-api==1.0.0

